add control date mapping
